requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
PyYAML==6.0.2
requests==2.32.5
ruff==0.8.4
urllib3==2.6.2
Werkzeug==3.1.4
